Add migration to bigint columns

+++ lib/Migration/Version2060Date20200302131958.php
@@ -26,7 +26,7 @@ class Version2060Date20200302131958 extends SimpleMigrationStep {
 
 		if (!$schema->hasTable('richdocuments_wopi')) {
 			$table = $schema->createTable('richdocuments_wopi');
-			$table->addColumn('id', 'integer', [
+			$table->addColumn('id', 'bigint', [
 				'autoincrement' => true,
 				'notnull' => true,
 				'length' => 20,
@@ -44,11 +44,11 @@ class Version2060Date20200302131958 extends SimpleMigrationStep {
 				'notnull' => false,
 				'length' => 64,
 			]);
-			$table->addColumn('fileid', 'integer', [
+			$table->addColumn('fileid', 'bigint', [
 				'notnull' => true,
 				'length' => 20,
 			]);
-			$table->addColumn('version', 'integer', [
+			$table->addColumn('version', 'bigint', [
 				//'notnull' => true,
 				'notnull' => false,
 				'length' => 20,
@@ -69,16 +69,16 @@ class Version2060Date20200302131958 extends SimpleMigrationStep {
 				'length' => 32,
 				'default' => '',
 			]);
-			$table->addColumn('expiry', 'integer', [
+			$table->addColumn('expiry', 'bigint', [
 				'notnull' => false,
 				'length' => 20,
 				'unsigned' => true,
 			]);
-			$table->addColumn('template_destination', 'integer', [
+			$table->addColumn('template_destination', 'bigint', [
 				'notnull' => false,
 				'length' => 20,
 			]);
-			$table->addColumn('template_id', 'integer', [
+			$table->addColumn('template_id', 'bigint', [
 				'notnull' => false,
 				'length' => 20,
 			]);
@@ -114,7 +114,7 @@ class Version2060Date20200302131958 extends SimpleMigrationStep {
 
 		if (!$schema->hasTable('richdocuments_direct')) {
 			$table = $schema->createTable('richdocuments_direct');
-			$table->addColumn('id', 'integer', [
+			$table->addColumn('id', 'bigint', [
 				'autoincrement' => true,
 				'notnull' => true,
 				'length' => 20,
@@ -128,22 +128,22 @@ class Version2060Date20200302131958 extends SimpleMigrationStep {
 				'notnull' => false,
 				'length' => 64,
 			]);
-			$table->addColumn('fileid', 'integer', [
+			$table->addColumn('fileid', 'bigint', [
 				'notnull' => true,
 				'length' => 20,
 			]);
-			$table->addColumn('timestamp', 'integer', [
+			$table->addColumn('timestamp', 'bigint', [
 				// 'notnull' => true,
 				'notnull' => false,
 				'length' => 20,
 				'default' => 0,
 				'unsigned' => true,
 			]);
-			$table->addColumn('template_destination', 'integer', [
+			$table->addColumn('template_destination', 'bigint', [
 				'notnull' => false,
 				'length' => 20,
 			]);
-			$table->addColumn('template_id', 'integer', [
+			$table->addColumn('template_id', 'bigint', [
 				'notnull' => false,
 				'length' => 20,
 			]);
@@ -154,7 +154,7 @@ class Version2060Date20200302131958 extends SimpleMigrationStep {
 
 		if (!$schema->hasTable('richdocuments_assets')) {
 			$table = $schema->createTable('richdocuments_assets');
-			$table->addColumn('id', 'integer', [
+			$table->addColumn('id', 'bigint', [
 				'autoincrement' => true,
 				'notnull' => true,
 				'length' => 20,
@@ -164,7 +164,7 @@ class Version2060Date20200302131958 extends SimpleMigrationStep {
 				'notnull' => false,
 				'length' => 64,
 			]);
-			$table->addColumn('fileid', 'integer', [
+			$table->addColumn('fileid', 'bigint', [
 				'notnull' => true,
 				'length' => 20,
 			]);
@@ -172,7 +172,7 @@ class Version2060Date20200302131958 extends SimpleMigrationStep {
 				'notnull' => false,
 				'length' => 64,
 			]);
-			$table->addColumn('timestamp', 'integer', [
+			$table->addColumn('timestamp', 'bigint', [
 				// 'notnull' => true,
 				'notnull' => false,
 				'length' => 20,
