+++ lib/Migration/Version30704Date20200626072306.php
@@ -24,7 +24,7 @@ class Version30704Date20200626072306 extends SimpleMigrationStep {
 		$table = $schema->getTable('richdocuments_wopi');
 
 		if (!$table->hasColumn('template_id')) {
-			$table->addColumn('template_id', 'integer', [
+			$table->addColumn('template_id', 'bigint', [
 				'notnull' => false,
 				'length' => 20,
 			]);
